[~] Moved  extension [~] Better pattern building #2600
